+++ packages/next-common/components/collective/metadata/index.js
@@ -3,6 +3,11 @@ import capitalize from "../../../utils/capitalize";
 import Proposal from "../../proposal";
 import KVList from "../../listInfo/kvList";
 import User from "../../user";
+import { useChain } from "../../../context/chain";
+import MarketMetadata from "./marketMetadata";
+import Chains from "../../../utils/consts/chains";
+import { useDetailType } from "../../../context/page";
+import { detailPageCategory } from "../../../utils/consts/business/category";
 
 const keys = {
   proposer: "proposer",
@@ -18,7 +23,12 @@ export default function CollectiveMetadata({
   hash,
   call,
   index,
+  marketId,
+  marketMetadata,
 }) {
+  const chain = useChain();
+  const detailType = useDetailType();
+
   const proposerItem = [
     capitalize(keys.proposer),
     <User add={proposer} fontSize={14} key="proposer" />,
@@ -38,6 +48,16 @@ export default function CollectiveMetadata({
 
   if (call) {
     data.push([<Proposal call={call} key="proposal" />]);
+  }
+
+  if (chain === Chains.zeitgeist && detailPageCategory.ADVISORY_MOTION === detailType) {
+    data.push([
+      <MarketMetadata
+        key={"marketMetadata"}
+        id={marketId}
+        metadata={marketMetadata}
+      />,
+    ]);
   }
 
   return <KVList title="Metadata" data={data} />;