+++ packages/next-common/components/linkPost/styled.js
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+import { p_14_normal } from "../../styles/componentCss";
+
+export const Info = styled.div`
+  padding: 10px 16px;
+  background: #f6f7fa;
+  border-radius: 4px;
+
+  ${p_14_normal}
+  color: #506176;
+`;
+
+export const ButtonWrapper = styled.div`
+  display: flex;
+  justify-content: flex-end;
+`;