+++ packages/next-common/components/post/postEdit.js
@@ -13,6 +13,7 @@ import { EditablePanel } from "../styled/panel";
 import { usePost, usePostTitle } from "../../context/post";
 import { useDetailType } from "../../context/page";
 import PostLabel from "./postLabel";
+import { detailPageCategory } from "../../utils/consts/business/category";
 
 const Wrapper = styled(EditablePanel)`
   textarea:read-only,
@@ -41,7 +42,8 @@ export default function PostEdit({ setIsEdit, updatePost }) {
   const [title, setTitle] = useState(defaultTitle);
   const [updating, setUpdating] = useState(false);
   const [bannerCid, setBannerCid] = useState(post.bannerCid);
-  const [selectedLabels, setSelectedLabels] = useState(post.labels);
+  const [selectedLabels, setSelectedLabels] = useState(post.labels || []);
+  const postType = useDetailType();
 
   const [isSetBanner, setIsSetBanner] = useState(!!post.bannerCid);
   useEffect(() => {
@@ -93,10 +95,12 @@ export default function PostEdit({ setIsEdit, updatePost }) {
         </UploaderWrapper>
       )}
 
-      <PostLabel
-        selectedLabels={selectedLabels}
-        setSelectedLabels={setSelectedLabels}
-      />
+      {postType === detailPageCategory.POST && (
+        <PostLabel
+          selectedLabels={selectedLabels}
+          setSelectedLabels={setSelectedLabels}
+        />
+      )}
 
       <LabelWrapper>
         <Label>Issue</Label>