+++ packages/next-common/utils/hooks/useAcceptCookies.js
@@ -2,10 +2,7 @@ import { CACHE_KEY } from "../../utils/constants";
 import { useCookieValue } from "./useCookieValue";
 
 export function useAcceptCookies() {
-  const [accepted, setAccepted] = useCookieValue(
-    CACHE_KEY.acceptCookies,
-    false,
-  );
+  const [accepted, setAccepted] = useCookieValue(CACHE_KEY.acceptCookies);
 
   return [accepted, setAccepted];
 }