+++ packages/next/pages/post/[id].js
@@ -76,7 +76,7 @@ export const getServerSideProps = withLoginUser(async (context) => {
   ]);
 
   if (!detail) {
-    return to404(context);
+    return to404();
   }
 
   const postId = detail._id;