+++ packages/next/utils/viewfuncs/index.js
@@ -1,3 +1,4 @@
+import flatten from "lodash.flatten";
 import { addressEllipsis } from "next-common/utils";
 import { getMotionId } from "next-common/utils/motion";
 import { getTitle } from "next-common/utils/post";
@@ -6,6 +7,31 @@ import {
   childBountyBaseUrl,
 } from "next-common/utils/postBaseUrl";
 import { getPostLastActivityAt } from "next-common/utils/viewfuncs/postUpdatedTime";
+
+export const convertPolkassemblyReactionV2 = (chain, paReactions) =>
+  flatten(
+    Object.entries(paReactions || {}).map(([r, { usernames } = {}]) => usernames?.map(u => [r, u]))
+  ).map(([r, u]) => ({
+    reaction: r === "👍" ? 1 : 0,
+    user: u,
+  }));
+
+export const convertPolkassemblyCommentV2 = (chain, comment) => ({
+  reactions: convertPolkassemblyReactionV2(comment.comment_reactions),
+  id: comment.id,
+  content: comment.content,
+  createdAt: comment.created_at,
+  updatedAt: comment.updated_at,
+  author: {
+    username: comment.username,
+    address: comment.proposer,
+  },
+});
+
+export const toPolkassemblyCommentListItemV2 = (chain, item) => ({
+  ...convertPolkassemblyCommentV2(chain, item),
+  replies: item.replies?.map((r) => convertPolkassemblyCommentV2(chain, r)),
+});
 
 export const convertPolkassemblyUser = (chain, paUser) =>
   paUser?.[`${chain}_default_address`]