+++ CHANGELOG.md
@@ -2,11 +2,20 @@
 
 All notable changes to this project will be documented in this file.
 
+# 2.9.8
+date: 03/13/2023
+- Add login user profile link to the dropdown menu.
+- Customize zeitgeist advisory committee motion market info.
+- Improve popup responsive style.
+- Fix delegation items responsive scroll style.
+- Improve husky hooks for eslint.
+
 # 2.9.7
-date 03/03/2023
+date: 03/03/2023
 - Enable polkadot collectives.
 - Fix cmdk mask background color.
 - Improve eslint config and fix lint warnings.
+- Add warning bar for scam proposals.
 
 # 2.9.6